requirements.txt
asgiref==3.3.4
attrs==21.4.0
autopep8==1.6.0
boto3==1.21.42
botocore==1.24.42
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.0
django-allauth==0.44.0
django-annoying==0.10.6
django-cors-headers==3.7.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-extensions==3.1.5
django-model-utils==4.2.0
django-rest-auth==0.9.5
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
firebase-admin==5.0.0
google-api-core==1.31.5
google-api-python-client==2.42.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
httplib2==0.20.4
idna==3.3
iniconfig==1.1.1
jmespath==1.0.0
jsonfield==3.1.0
msgpack==1.0.3
mysql-connector-python==8.0.28
oauthlib==3.2.0
packaging==21.3
Pillow==8.3.2
pluggy==1.0.0
proto-plus==1.20.3
protobuf==3.19.4

py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.7
pytest==7.1.1
pytest-django==3.9.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==5.2.0
